pyproject.toml
[tool.poetry]
name = "inventory_control"
version = "0.1.0"
description = ""
authors = ["NatanCervinski <natanscer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
python-dotenv = "^0.15.0"
dynaconf = "^3.1.3"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.7.0"
mysqlclient = "^2.0.3"
Flask-RESTful = "^0.3.8"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.24.2"
Flask-JWT-Extended = "^4.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-flask = "^1.2.0"
black = "^20.8b1"
flake8 = "^3.9.1"
pytest-sugar = "^0.9.4"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
